update!
Changed hood angle from 25 to 30. uhhh added extra capability and tinkered with controls and such ...

think I messed with how the lift angle is set a little bit ... anwyways there might be some messups in here that need to be cleaned up but lincoln and anna were driving with this.